newest version from Alex now with correct BCs
